requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
langchain-text-splitters
pandas
langchain-groq
streamlit